// src/game.types.ts
export type ExpansionName =
  | "Base Game"
  | "Dead of Night"
  | "Under Dark Waves"
  | "Secrets of the Order"
  | "Recursive Echoes";

export type ScenarioName =
  | "Approach of Azathoth"
  | "Feast of Umordhoth"
  | "Veil of Twilight"
  | "Echoes of the Deep"
  | "Shots in the Dark"
  | "Silence of Tsathogua"
  | "The Pale Lantern"
  | "Tyrants of Ruin"
  | "Dreams of R'lyeh"
  | "Ithaqua's Children"
  | "Bound to Serve"
  | "The Dead Cry Out"
  | "The Key and the Gate";